requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
idna==3.7
Markdown==3.5.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.1
